# api.yaml
openapi: 3.0.4
info:
  title: Bilheteira AFC Coimbra
  description: |-
    API para a venda de bilhetes para a Divisão de Elite da Associação de Futebol de Coimbra
  contact:
    email: afcoimbra@afcoimbra.com
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: Servidor local
...
security:
  - bearerAuth: []
paths:
...
  /client:
    get:
      tags: [Cliente]
      summary: Todos os clientes
      operationId: getAllClients
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cliente'
        '500':
          description: Falha no servidor
    post:
      tags: [Cliente]
      summary: Cria um novo cliente
      operationId: createClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteInput'
      responses:
        '201':
          description: Cliente criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '400':
          description: Dados inválidos
        '500':
          description: Falha no servidor
  /client/{id_cliente}:
    parameters:
      - name: id_cliente
        in: path
        required: true
        description: ID do cliente
        schema:
          type: integer
          format: int64
          example: 1
    get:
      tags: [Cliente]
      summary: Encontrar cliente pelo ID
      operationId: getClientById
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '404':
          description: Cliente não encontrado
        '500':
          description: Falha no servidor
    put:
      tags: [Cliente]
      summary: Atualiza um cliente pelo ID
      operationId: updateClientById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteInput'
      responses:
        '200':
          description: Cliente atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '400':
          description: Dados inválidos
        '404':
          description: Cliente não encontrado
        '500':
          description: Falha no servidor
    delete:
      tags: [Cliente]
      summary: Apaga um cliente pelo ID
      operationId: deleteClient
      responses:
        '204':
          description: Cliente eliminado com sucesso
        '404':
          description: Cliente não encontrado
        '500':
          description: Falha no servidor
...
  /bilhete:
    get:
      tags: [Bilhete]
      summary: Todos os bilhetes
      operationId: getAllBilhetes
      parameters:
        - name: id_jogo
          in: query
          required: false
          description: Filtrar bilhetes por jogo
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Lista de bilhetes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bilhete'
        '500':
          description: Falha no servidor
    post:
      tags: [Bilhete]
      summary: Cria um novo bilhete
      operationId: createBilhete
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Bilhete'
      responses:
        '201':
          description: Bilhete criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bilhete'
        '400':
          description: Dados inválidos
        '500':
          description: Falha no servidor
  /bilhete/{id_bilhete}:
    parameters:
      - name: id_bilhete
        in: path
        required: true
        description: ID do bilhete
        schema:
          type: integer
          format: int64
          example: 1
    get:
      tags: [Bilhete]
      summary: Encontrar bilhete pelo ID
      operationId: getBilheteById
      responses:
        '200':
          description: Bilhete encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bilhete'
        '404':
          description: Bilhete não encontrado
        '500':
          description: Falha no servidor
    put:
      tags: [Bilhete]
      summary: Atualiza um bilhete pelo ID
      operationId: updateBilhete
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Bilhete'
      responses:
        '200':
          description: Bilhete atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bilhete'
        '400':
          description: Dados inválidos
        '404':
          description: Bilhete não encontrado
        '500':
          description: Falha no servidor
    delete:
      tags: [Bilhete]
      summary: Apaga um bilhete pelo ID
      operationId: deleteBilhete
      responses:
        '204':
          description: Bilhete eliminado com sucesso
        '404':
          description: Bilhete não encontrado
        '500':
          description: Falha no servidor
  /compra:
    get:
      tags: [Compra]
      summary: Todas as compras
      operationId: getAllCompras
      responses:
        '200':
          description: Lista de compras
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Compra'
        '500':
          description: Falha no servidor
    post:
      tags: [Compra]
      summary: Efetua uma nova compra
      operationId: createCompra
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Compra'
      responses:
        '201':
          description: Compra realizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Compra'
        '400':
          description: Dados inválidos
        '500':
          description: Falha no servidor
  /compra/{id_compra}:
    parameters:
      - name: id_compra
        in: path
        required: true
        description: ID da compra
        schema:
          type: integer
          format: int64
          example: 1
    get:
      tags: [Compra]
      summary: Encontrar compra pelo ID
      operationId: getCompraById
      responses:
        '200':
          description: Compra encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Compra'
        '404':
          description: Compra não encontrada
        '500':
          description: Falha no servidor
    put:
      tags: [Compra]
      summary: Atualiza o estado de uma compra
      operationId: updateCompra
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                estado:
                  type: string
                  enum: [pago, por pagar, cancelado]
                  example: pago
      responses:
        '200':
          description: Compra atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Compra'
        '400':
          description: Dados inválidos
        '404':
          description: Compra não encontrada
        '500':
          description: Falha no servidor
    delete:
      tags: [Compra]
      summary: Apaga uma compra pelo ID
      operationId: deleteCompra
      responses:
        '204':
          description: Compra eliminada com sucesso
        '404':
          description: Compra não encontrada
        '500':
          description: Falha no servidor
  /compra/cliente/{id_cliente}:
    parameters:
      - name: id_cliente
        in: path
        required: true
        description: ID do cliente
        schema:
          type: integer
          format: int64
          example: 1
    get:
      tags: [Compra]
      summary: Todas as compras de um cliente específico
      operationId: getComprasByCliente
      responses:
        '200':
          description: Lista de compras do cliente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Compra'
        '404':
          description: Cliente não encontrado
        '500':
          description: Falha no servidor
components:
  schemas:
...
    Cliente:
      type: object
      properties:
        id_cliente:
          type: integer
          format: int64
          description: Chave primária da entidade Cliente
          example: 1
        nome:
          type: string
          description: Nome completo do Cliente
          example: Carlos Silva
        email:
          type: string
          description: Email do Cliente
          example: email@gmail.com
        nif:
          type: integer
          format: int64
          description: NIF do Cliente
          example: 123456789
        telemovel:
          type: integer
          format: int64
          description: Telemóvel do Cliente
          example: 931393191
        dta_nascimento:
          type: string
          format: date
          description: Data de nascimento do cliente
          example: "2002-08-10"

    # Schema de input (criação/atualização — sem id)
    ClienteInput:
      type: object
      required:
        - nome
        - email
        - nif
      properties:
        nome:
          type: string
          description: Nome completo do Cliente
          example: Carlos Silva
        email:
          type: string
          description: Email do Cliente
          example: email@gmail.com
        password:
          type: string
          description: Password do Cliente
          example: password123
        nif:
          type: integer
          format: int64
          description: NIF do Cliente
          example: 123456789
        telemovel:
          type: integer
          format: int64
          description: Telemóvel do Cliente
          example: 931393191
        dta_nascimento:
          type: string
          format: date
          description: Data de nascimento do cliente
          example: "2002-08-10"
...
    Bilhete:
      type: object
      required:
        - custo
        - id_jogo
      properties:
        id_bilhete:
          type: integer
          format: int64
          description: Chave primária da entidade Bilhete
          example: 1
        custo:
          type: number
          format: float
          description: Custo do bilhete
          example: 4.50
        id_jogo:
          type: integer
          format: int64
          description: ID do jogo ao qual o bilhete pertence
          example: 1
    Compra:
      type: object
      required:
        - id_cliente
        - id_bilhete
      properties:
        id_compra:
          type: integer
          format: int64
          description: Chave primária da entidade Compra
          example: 1
        estado:
          type: string
          enum: [pago, por pagar, cancelado]
          description: Estado em que se encontra a compra
          example: pago
        dta_compra:
          type: string
          format: date-time
          description: Data e hora da compra
          example: "1999-09-09T16:50:00Z"
        id_cliente:
          type: integer
          format: int64
          description: ID do cliente que efetuou a compra
          example: 1
        id_bilhete:
          type: integer
          format: int64
          description: ID do bilhete adquirido
          example: 1
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT